+++ cognite/client/data_classes/data_modeling/__init__.py
@@ -82,9 +82,9 @@ from cognite.client.data_classes.data_modeling.spaces import Space, SpaceApply, 
 from cognite.client.data_classes.data_modeling.typed_instances import (
     PropertyOptions,
     TypedEdge,
-    TypedEdgeWrite,
+    TypedEdgeApply,
     TypedNode,
-    TypedNodeWrite,
+    TypedNodeApply,
 )
 from cognite.client.data_classes.data_modeling.views import (
     ConnectionDefinition,
@@ -197,8 +197,8 @@ __all__ = [
     "InstanceApply",
     "query",
     "PropertyOptions",
-    "TypedEdgeWrite",
-    "TypedNodeWrite",
+    "TypedEdgeApply",
+    "TypedNodeApply",
     "TypedNode",
     "TypedEdge",
 ]